ISO/TC307/WG02 - Identity;  Blockchain and distributed ledger technologies — Overview of Identity
Overview of Identity - an identity system must accommodate individuals, legal individuals, ‘things’ and processes; so, the framework conceptualizes an entity to be an individual, legal entity, a thing or a process. Identity defines the whole set of attributes that describe an entity and determines the types of transactions in which it can participate in.